requirements.txt
asgiref==3.5.2
CacheControl==0.12.11
cachetools==5.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==37.0.2
defusedxml==0.7.1
distlib==0.3.4
Django==3.2
django-allauth==0.51.0
django-cors-headers==3.13.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
dlib==19.24.0
face-recognition==1.3.0
face-recognition-models==0.3.0
filelock==3.7.0
firebase-admin==5.2.0
google-api-core==2.8.1
google-api-python-client==2.49.0
google-auth==2.6.6
google-auth-httplib2==0.1.0
google-cloud-core==2.3.0
google-cloud-firestore==2.5.2
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.2
grpcio==1.46.3
grpcio-status==1.46.3
httplib2==0.20.4
idna==3.3
imutils==0.5.4
msgpack==1.0.4
numpy==1.22.2
oauthlib==3.2.0
opencv-contrib-python==4.5.5.64
opencv-python==4.5.5.64
pandas==1.4.1
Pillow==9.1.1
platformdirs==2.5.2
proto-plus==1.20.5
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
PyQt3D==5.15.5
PyQt5==5.15.6
PyQt5-sip==12.9.0
PyQtChart==5.15.5
PyQtDataVisualization==5.15.5
PyQtNetworkAuth==5.15.5
PyQtPurchasing==5.15.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
sqlparse==0.4.2
TBB==0.2
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
